Use Eigen::aligned_allocator to fix 32bit align
Attempts to fix 32bit alignment error in Eigen related code by
using the Eigen::aligned_allocator in all STL Containers with Eigen.

See these resources for more information than I understand:

* http://eigen.tuxfamily.org/dox-devel/group__TopicStlContainers.html

* (Cause 2:) http://eigen.tuxfamily.org/dox-devel/group__TopicUnalignedArrayAssert.html

Note: This is not a completely informed fix ... it just seemed to
finally fix the alignment error and make the program not crash
on 32-bit 12.04 with groovy (no kinect).