[pose controller] make sure to wrap angles and proper atan2 control rule.
Angles need to be wrapped on -pi, pi otherwise the control
rules generated from these will blow up (they are proportional
to the angle). Do that in this class so the user can't
inadvertantly make bugs.

There was a atan -> atan2 update in 0cdbb765a which didn't implement
atan2 correctly, this fixes that. We did not see the problem for a long
time simply because the divisor was 1.0 which is the same as multiplying
by 1.0.